requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-UUID==0.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.11.4
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
